+++ src/pages/BorrowerProfile/BorrowerProfile.vue
@@ -1,7 +1,6 @@
 <template>
 	<www-page
 		id="borrower-profile"
-		:header-theme="headerTheme"
 		main-class="kv-tailwind"
 	>
 		<article class="tw-relative tw-bg-secondary">
@@ -57,7 +56,6 @@
 </template>
 
 <script>
-import { lightHeader } from '@/util/siteThemes';
 
 import WwwPage from '@/components/WwwFrame/WwwPage';
 import ContentContainer from '@/components/BorrowerProfile/ContentContainer';
@@ -90,7 +88,6 @@ export default {
 	},
 	data() {
 		return {
-			headerTheme: lightHeader,
 			loanId: Number(this.$route.params.id || 0),
 			showLenders: true,
 			showTeams: true,